Avoid setting absent values in MZFayeMessage
This allows us to later determine whether the value was present or absent in the original message from the server.

Specifically, this fixes a problem where 'successful' was getting set to @NO when it was absent from the original message. There is significance in the difference between the server sending this value as 'false', or omitting it entirely; the former means it is a negative response to a message that I attempted to publish, but the latter means it is not a response at all.